# docs/swagger.yaml
components:
  schemas:
...
    api.authorResponse:
      description: 作者信息
      properties:
        avatar_url:
          description: 头像地址
          type: string
        id:
          description: 用户 ID
          type: string
        username:
          description: 用户名
          type: string
        viewer_following:
          description: 当前用户是否已关注作者
          type: boolean
      required:
      - id
      - viewer_following
      type: object
...
    api.collectPostRequest:
      properties:
        folder_id:
          description: 收藏夹 ID,可选
          type: string
      type: object
    api.collectionStatusResponse:
      properties:
        collect_count:
          description: 收藏数量
          minimum: 0
          type: integer
        viewer_collected:
          description: 是否已收藏
          type: boolean
      required:
      - collect_count
      - viewer_collected
      type: object
...
    api.createFolderRequest:
      properties:
        name:
          description: 收藏夹名称
          maxLength: 50
          type: string
      required:
      - name
      type: object
...
    api.folderResponse:
      properties:
        created_at:
          description: 创建时间
          type: string
        id:
          description: 收藏夹 ID
          type: string
        name:
          description: 收藏夹名称
          type: string
      type: object
...
    api.listPostsItemResponse:
      properties:
        author:
          $ref: '#/components/schemas/api.authorResponse'
        collect_count:
          description: 收藏数
          minimum: 0
          type: integer
        comment_count:
          description: 评论数
          minimum: 0
          type: integer
        content:
          description: 内容
          type: string
        cover_url:
          description: 封面 URL
          type: string
        created_at:
          description: 创建时间
          type: string
        id:
          description: 帖子 ID
          type: string
        like_count:
          description: 点赞数
          minimum: 0
          type: integer
        title:
          description: 标题
          type: string
        view_count:
          description: 浏览量
          minimum: 0
          type: integer
        viewer_collected:
          description: 当前用户是否已收藏
          type: boolean
        viewer_liked:
          description: 当前用户是否已点赞
          type: boolean
      required:
      - collect_count
      - comment_count
      - id
      - like_count
      - view_count
      - viewer_collected
      - viewer_liked
      type: object
...
    api.pageResponse-api_folderResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/api.folderResponse'
          type: array
          uniqueItems: false
        page:
          minimum: 1
          type: integer
        page_size:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      type: object
...
    api.pageResponse-api_listPostsItemResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/api.listPostsItemResponse'
          type: array
          uniqueItems: false
        page:
          minimum: 1
          type: integer
        page_size:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      type: object
...
    api.updateFolderRequest:
      properties:
        name:
          description: 收藏夹名称
          maxLength: 50
          type: string
      required:
      - name
      type: object
...
    render.Response-api_collectionStatusResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.collectionStatusResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_folderResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.folderResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_pageResponse-api_folderResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.pageResponse-api_folderResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_pageResponse-api_listPostsItemResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.pageResponse-api_listPostsItemResponse'
        msg:
          type: string
      type: object
...
    render.ResponseWithoutData:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /me/collections:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_pageResponse-api_listPostsItemResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 获取我的收藏列表
      tags:
      - collections
  /me/collections/folders:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_pageResponse-api_folderResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 获取收藏夹列表
      tags:
      - collections
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.createFolderRequest'
              description: 收藏夹信息
              summary: body
        description: 收藏夹信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_folderResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建收藏夹
      tags:
      - collections
  /me/collections/folders/{folder_id}:
    delete:
      parameters:
      - description: 收藏夹 ID
        in: path
        name: folder_id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.ResponseWithoutData'
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 删除收藏夹
      tags:
      - collections
    patch:
      parameters:
      - description: 收藏夹 ID
        in: path
        name: folder_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.updateFolderRequest'
              description: 收藏夹信息
              summary: body
        description: 收藏夹信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_folderResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 修改收藏夹
      tags:
      - collections
...
  /posts/{post_id}/collection:
    delete:
      parameters:
      - description: 帖子 ID
        in: path
        name: post_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_collectionStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 取消收藏帖子
      tags:
      - collections
    put:
      parameters:
      - description: 帖子 ID
        in: path
        name: post_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.collectPostRequest'
              description: 收藏夹
              summary: body
        description: 收藏夹
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_collectionStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 收藏帖子
      tags:
      - collections
...
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1